# server/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Sonarly API
  version: 2.0.0
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: >
    Native REST API of the Sonarly Go server. Documented routes
    are the native /api surface plus the /health, /healthz and /ready
    liveness probes;
    the OpenSubsonic adapter under /rest is a separate (Subsonic-compatible)
    contract and is intentionally not part of this spec.

    Error contract: every non-2xx response body is `{"error": "..."}` with a
    generic message for 5xx (driver errors never reach the client).
servers:
  - url: /
# Default auth: the signed sessionId cookie OR the X-API-Key header.
# Public operations override with `security: []`; /api/events overrides with
# session-cookie-only (EventSource cannot set headers).
security:
  - sessionId: []
  - apiKey: []
...
paths:
...
  /api/ingest:
    get:
      operationId: listIngestJobs
      tags: [ingest]
      summary: Last 100 per-file ingest rows
      responses:
        '200':
          description: Newest-first ingest rows (empty array when none).
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items: { $ref: '#/components/schemas/IngestJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: wipeIngestJobs
      tags: [ingest]
      summary: Wipe the ingest history table (admin)
      responses:
        '200':
          description: Table wiped.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/ingest/{id}:
    get:
      operationId: getIngestJob
      tags: [ingest]
      summary: One ingest row
      parameters:
        - $ref: '#/components/parameters/IngestJobId'
      responses:
        '200':
          description: The ingest row.
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job: { $ref: '#/components/schemas/IngestJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteIngestJob
      tags: [ingest]
      summary: Delete one ingest row (admin)
      parameters:
        - $ref: '#/components/parameters/IngestJobId'
      responses:
        '200':
          description: Row deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/ingest/trigger:
    post:
      operationId: triggerIngest
      tags: [ingest]
      summary: Enqueue an ingest job for a library's drop dir (admin)
      description: >
        The body is optional: an empty/absent body (or `{"libraryId": ""}`)
        targets the default library. The library's drop dir is created if
        missing before the job is enqueued.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                libraryId: { type: string, format: uuid }
      responses:
        '200':
          description: Ingest job enqueued.
          content:
            application/json:
              schema:
                type: object
                required: [ok, jobId]
                properties:
                  ok: { type: boolean, enum: [true] }
                  jobId: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/conflicts:
    get:
      operationId: listConflicts
      tags: [ingest]
      summary: Songs parked on " (n)" collision paths (admin)
      responses:
        '200':
          description: Collision candidates with artist/album names.
          content:
            application/json:
              schema:
                type: object
                required: [conflicts]
                properties:
                  conflicts:
                    type: array
                    items: { $ref: '#/components/schemas/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteConflicts
      tags: [ingest]
      summary: Delete collision files and their song rows (admin)
      description: >
        For every collision song the FILE is deleted first and the row
        second, so a row pointing at a live file never outlives it.
      responses:
        '200':
          description: Number of conflicts deleted.
          content:
            application/json:
              schema:
                type: object
                required: [ok, deleted]
                properties:
                  ok: { type: boolean, enum: [true] }
                  deleted: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/settings/media:
    get:
      operationId: getMediaSettings
      tags: [ingest]
      summary: Media settings with pattern templates (admin)
      responses:
        '200':
          description: The effective settings plus the template presets.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      operationId: patchMediaSettings
      tags: [ingest]
      summary: Update media settings (admin)
      description: >
        Every field is optional; the pattern must be relative with no `..`
        segments, the strategy must be a known DuplicateStrategy, and the
        retention clamps to 1–365. Answers the effective settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizePattern: { type: string, minLength: 1 }
                duplicateStrategy: { $ref: '#/components/schemas/DuplicateStrategy' }
                reviewRetentionDays: { type: integer, minimum: 1, maximum: 365 }
      responses:
        '200':
          description: The effective settings after the patch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaSettingsWithoutTemplates' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/organize:
    post:
      operationId: organize
      tags: [ingest]
      summary: Enqueue an organize job (admin, 202)
      description: >
        Re-applies every library's organize pattern to the files already in
        the library. The job runs on the worker — never inline on the
        request goroutine.
      responses:
        '202':
          description: Job enqueued.
          content:
            application/json:
              schema:
                type: object
                required: [jobId]
                properties:
                  jobId: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/organize/job:
    post:
      operationId: organizeJob
      tags: [ingest]
      summary: Enqueue an organize job (admin, 200 — legacy response shape)
      responses:
        '200':
          description: Job enqueued.
          content:
            application/json:
              schema:
                type: object
                required: [jobId]
                properties:
                  jobId: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/organize/status/{jobId}:
    get:
      operationId: organizeStatus
      tags: [ingest]
      summary: One organize job's status (admin)
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: The job status document.
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job: { $ref: '#/components/schemas/OrganizeJobStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    sessionId:
      type: apiKey
      in: cookie
      name: sessionId
      description: Signed session cookie (HttpOnly, SameSite=Strict), set by /api/login and /api/setup. The only credential /api/events accepts.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative to the session cookie on ordinary /api routes (not /api/events). SHA-256 digest looked up in api_keys.
  parameters:
...
    IngestJobId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Ingest job id (UUID).
...
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string }
      description: Scan job id (UUID).
  responses:
    BadRequest:
      description: Malformed input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed (admin-gated route, or playlist access denied).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or out of the caller's scope).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    # --- enums ---------------------------------------------------------------
...
    DuplicateStrategy:
      type: string
      enum:
        - replace_file_and_metadata
        - keep_file_replace_metadata
        - replace_file_aggregate_metadata
        - keep_file_aggregate_metadata
        - skip
...
    JobStatusName:
      type: string
      enum: [pending, running, completed, failed]
    # --- libraries -------------------------------------------------------------
...
    MediaSettings:
      type: object
      required: [organizePattern, duplicateStrategy, reviewRetentionDays, templates]
      properties:
        organizePattern: { type: string }
        duplicateStrategy: { $ref: '#/components/schemas/DuplicateStrategy' }
        reviewRetentionDays: { type: integer, minimum: 1, maximum: 365 }
        templates:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: string }
    MediaSettingsWithoutTemplates:
      type: object
      required: [organizePattern, duplicateStrategy, reviewRetentionDays]
      properties:
        organizePattern: { type: string }
        duplicateStrategy: { $ref: '#/components/schemas/DuplicateStrategy' }
        reviewRetentionDays: { type: integer, minimum: 1, maximum: 365 }
    OrganizeJobStatus:
      type: object
      required: [id, type, status]
      properties:
        id: { type: string }
        type: { type: string, enum: [organize] }
        status: { $ref: '#/components/schemas/JobStatusName' }
        startedAt: { type: string }
        finishedAt: { type: string }
        stats: { type: object, additionalProperties: true }
    # --- users ---------------------------------------------------------------
...
    IngestJob:
      type: object
      required: [id, runId, sourcePath, status, duplicate, createdAt, updatedAt]
      properties:
        id: { type: string }
        runId: { type: string }
        sourcePath: { type: string }
        status: { type: string, enum: [pending, needs_review, imported, skipped, failed] }
        targetPath: { type: [string, 'null'] }
        error: { type: [string, 'null'] }
        duplicate: { type: boolean }
        duplicateStrategy:
          anyOf:
            - { $ref: '#/components/schemas/DuplicateStrategy' }
            - { type: 'null' }
        createdAt: { type: string }
        updatedAt: { type: string }
    Conflict:
      type: object
      description: One collision candidate — an active song parked on a ` (n)` collision path.
      required: [id, filePath, title]
      properties:
        id: { type: string }
        filePath: { type: string }
        title: { type: string }
        artistName: { type: string }
        albumName: { type: string }